HW3.Page objects. Extract Superclass refactoring.
